@@ -8,24 +8,26 @@
   },
   "dependencies": {
     "@rails/ujs": "^6.0.0",
-    "@sentry/browser": "^4.5.3",
-    "@sentry/react": "^6.18.2",
-    "@types/react": "^16.4.18",
-    "@types/react-dom": "^16.0.9",
+    "@sentry/browser": "^10.25.0",
+    "@sentry/react": "^10.25.0",
+    "@types/react": "^19.2.5",
+    "@types/react-dom": "^19.2.3",
     "axios": "^1.13.2",
     "bootstrap": "^4.1.3",
     "font-awesome": "^4.7.0",
     "jquery": "^3.3.1",
     "popper.js": "^1.14.4",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
     "sass": "^1.94.0",
     "typescript": "^5.9.3",
     "util": "^0.12.4"
   },
   "devDependencies": {
+    "@types/jquery": "^3.5.33",
     "prettier": "^3.1.1",
     "vite": "^7.2.2",
+    "vite-plugin-checker": "^0.11.0",
     "vite-plugin-full-reload": "^1.2.0",
     "vite-plugin-ruby": "^5.1.0"
   }
